feat(added): Added and fixed code design
