+++ pubsub-debug.log
@@ -1,12 +1,10 @@
 This is the Google Pub/Sub fake.
 Implementation may be incomplete or differ from the real system.
-Mar 04, 2021 1:51:43 PM com.google.cloud.pubsub.testing.v1.Main main
+Mar 05, 2021 10:50:31 AM com.google.cloud.pubsub.testing.v1.Main main
 INFO: IAM integration is disabled. IAM policy methods and ACL checks are not supported
-Mar 04, 2021 1:51:44 PM io.gapi.emulators.netty.NettyUtil applyJava7LongHostnameWorkaround
-INFO: Applied Java 7 long hostname workaround.
-Mar 04, 2021 1:51:44 PM com.google.cloud.pubsub.testing.v1.Main main
+Mar 05, 2021 10:50:33 AM io.gapi.emulators.netty.NettyUtil applyJava7LongHostnameWorkaround
+INFO: Unable to apply Java 7 long hostname workaround.
+Mar 05, 2021 10:50:33 AM com.google.cloud.pubsub.testing.v1.Main main
 INFO: Server started, listening on 9103
-Mar 04, 2021 1:51:44 PM io.gapi.emulators.grpc.GrpcServer$3 operationComplete
+Mar 05, 2021 10:50:34 AM io.gapi.emulators.grpc.GrpcServer$3 operationComplete
 INFO: Adding handler(s) to newly registered Channel.
-*** shutting down gRPC server since JVM is shutting down
-*** server shut down
