+++ src/app/app-routing.module.ts
@@ -1,13 +1,13 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
+import { InventoryItemsComponent } from './body/inventory-items/inventory-items.component';
 import { InventoryComponent } from './body/inventory/inventory.component';
-import { ItemsComponent } from './body/items/items.component';
 import { LoginComponent } from './body/login/login.component';
 
 const routes: Routes = [
   { path: "login", component: LoginComponent },
   { path: "Inventory", component: InventoryComponent },
-  { path: "Inventory/:collection", component: ItemsComponent },
+  { path: "Inventory/Items", component: InventoryItemsComponent },
 ];
 
 @NgModule({
