+++ src/app/body/inventory-items/inventory-items.component.css
@@ -1,0 +1,3 @@
+table {
+    font-size: 12px;
+}
