+++ src/app/body/inventory-items/inventory-items.component.spec.ts
@@ -1,20 +1,20 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { ItemsComponent } from './items.component';
+import { InventoryItemsComponent } from './inventory-items.component';
 
-describe('ItemsComponent', () => {
-  let component: ItemsComponent;
-  let fixture: ComponentFixture<ItemsComponent>;
+describe('InventoryItemsComponent', () => {
+  let component: InventoryItemsComponent;
+  let fixture: ComponentFixture<InventoryItemsComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ ItemsComponent ]
+      declarations: [ InventoryItemsComponent ]
     })
     .compileComponents();
   });
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(ItemsComponent);
+    fixture = TestBed.createComponent(InventoryItemsComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
