+++ src/app/body/inventory/create-product/create-product.component.html
@@ -37,11 +37,11 @@
                     </div>
                 </div>
                 <div class="form-group row">
-                    <label for="description" class="col-4 col-form-label">Availability</label>
+                    <label for="description" class="col-4 col-form-label">Status</label>
                     <div class="col-8">
                         <select class="input" [(ngModel)]="productAvailability" [ngModelOptions]="{standalone: true}">
-                            <option>Yes</option>
-                            <option>No</option>
+                            <option>Available</option>
+                            <option>SoldOut</option>
                         </select>
                     </div>
                 </div>
