--- src/app/body/items/item-card/item-card.component.css
@@ -1,13 +1,0 @@
-.input,
-.input:focus {
-    border-radius: 10px;
-    padding: 10px;
-    border: 1px solid black;
-    outline: none;
-    width: 100%;
-}
-
-#button {
-    background-color: black;
-    color: white;
-}
