+++ ui-debug.log
@@ -1,1 +1,1 @@
-Web / API server started at http://127.0.0.1:9104
+Web / API server started at http://localhost:9104
